Introduce BackendService for managing external storage backends
Backends are registered to the BackendService through new data
structures:

BackendConfig specifies the basic information for a backend: class,
human-readable name, visibility to personal/admin, initial priority. It
takes an array of BackendParameter's to specifiy the parameters

BackendParameter stores a parameter configuration for an external
storage: name of parameter, human-readable name, type of parameter
(text, password, hidden, checkbox), flags (optional or not).

When a BackendConfig has checkDependencies() called on it, it will
return a (possibly empty) array of BackendDependency's that specify a
dependency of the backend. If an empty array is returned it means there
are no unhandled dependencies, so the backend can be used.

Storages in the StoragesController now get their parameters validated
server-side (fixes a TODO).